requirements.txt
openai-whisper
setuptools-rust
speechrecognition
openai
torch
google-cloud-speech
protobuf
loguru
mkdocs
mkdocs-material
pillow
pydub